+++ rxbindings/src/main/java/com/amplifyframework/rx/RxStorageCategoryBehavior.java
@@ -17,13 +17,18 @@ package com.amplifyframework.rx;
 
 import androidx.annotation.NonNull;
 
+import com.amplifyframework.core.Consumer;
+import com.amplifyframework.core.async.Cancelable;
 import com.amplifyframework.storage.StorageCategoryBehavior;
+import com.amplifyframework.storage.StorageException;
 import com.amplifyframework.storage.options.StorageDownloadFileOptions;
 import com.amplifyframework.storage.options.StorageListOptions;
 import com.amplifyframework.storage.options.StorageRemoveOptions;
 import com.amplifyframework.storage.options.StorageUploadFileOptions;
+import com.amplifyframework.storage.result.StorageDownloadFileResult;
 import com.amplifyframework.storage.result.StorageListResult;
 import com.amplifyframework.storage.result.StorageRemoveResult;
+import com.amplifyframework.storage.result.StorageTransferProgress;
 import com.amplifyframework.storage.result.StorageUploadFileResult;
 
 import java.io.File;
@@ -43,7 +48,7 @@ public interface RxStorageCategoryBehavior {
      *         download result or failure as a {@link Single}
      */
     @NonNull
-    RxStorageBinding.RxStorageDownloadOperation downloadFile(
+    RxStorageBinding.RxProgressAwareSingleOperation<StorageDownloadFileResult> downloadFile(
             @NonNull String key,
             @NonNull File local
     );
@@ -61,7 +66,7 @@ public interface RxStorageCategoryBehavior {
      *         by disposing the single subscription.
      */
     @NonNull
-    RxStorageBinding.RxStorageDownloadOperation downloadFile(
+    RxStorageBinding.RxProgressAwareSingleOperation<StorageDownloadFileResult> downloadFile(
             @NonNull String key,
             @NonNull File local,
             @NonNull StorageDownloadFileOptions options
@@ -148,4 +153,15 @@ public interface RxStorageCategoryBehavior {
             @NonNull String path,
             @NonNull StorageListOptions options
     );
+
+    /**
+     * Type alias that defines the generic parameters for a download operation.
+     * @param <T> The type that represents the result of a given operation.
+     * @see RxAdapters.RxProgressAwareCallbackMapper
+     */
+    interface RxStorageTransferCallbackMapper<T> {
+        Cancelable emitTo(Consumer<StorageTransferProgress> onProgress,
+                          Consumer<T> onItem,
+                          Consumer<StorageException> onError);
+    }
 }
